Provide support for FreeBSD
bash is not always located at /bin/bash so it's required to use /usr/bin/env to lookup for the correct bash location. On FreeBSD (and maybe on different operating systems) bind mount didn't work so nullfs need to be used on FreeBSD for example - providing an option in the conf-file. To support different remote locations it should be supported to specify the remote borg command path.